Update patch set 1
Patch Set 1:

Is there a good way to determine if someone is on a carrier that supports WB-AMR and is calling someone whose carrier also supports WB-AMR?

I'm wondering if this is WB-AMR related now.  Might explain why us AT&Ters aren't experiencing problems.  Sucks that our country is the "third world of mobile technology".

Patch-set: 1