Relocate static busybox to /sbin/static/busybox
Rename the static busybox that is being included in the boot image
to /sbin/static/busybox instead of /sbin/busybox. This change
prevents the static busybox from overriding the dynamic busybox
included with TWRP. This prevents errors when installing some
update zips that depend on the busybox included in recovery.

Patch set 1 of 2.

Change-Id: I2f3c14a85ecf7ab58ad55e413766e8e4b609ad95