Update patch set 2
Patch Set 2:

Max two changes because they are not really related. DEVICE_RESOLUTION is depricated as i understood from Dees_Troy and TW_THEME is the replacement. It is not really needed to make the recovery build.

The removal of the copying of the 1440x2560 resources in mkbootimg.mk is needed because it breaks the build. In recovey/gui/Android.mk the DEVICE_RESOLUTION values are mapped to the TW_THEME values used. This includes 1440x2560. But in mkbootimg.mk the actual resource files are copied and they are depricated and not included anymore in the repository. So the build fails.

Patch-set: 2
Reviewer: Gerrit User 1000049 <1000049@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0