Update patch set 4
Patch Set 4:

Your continued arguing about this shows that you have not understood what the problem is. 1) its an absolute no-go to change other people changes without their ok except it is an abandoned one. 2) if someone prefers a different order you have to accept it. 3) its just plain rude to continue to argue about your point. If you are unwilling to follow such basic rules for collaboration in this project you should consider to stop contributing to it. So you can count that as the first official warning to either start following our common principles or leave.

Patch-set: 4
Label: Code-Review=0